requirements.txt
selenium>=4.22
pytest>=8
pytest-html
webdriver-manager
python-dotenv
